use log2 from math.rs

+++ crates/starknet-os/src/hints/constants.rs
@@ -1,3 +1,3 @@
 // Compression constants
-pub const N_UNIQUE_VALUE_BUCKETS: u64 = 6;
-pub const TOTAL_N_BUCKETS: u64 = N_UNIQUE_VALUE_BUCKETS + 1;
+pub const N_UNIQUE_VALUE_BUCKETS: usize = 6;
+pub const TOTAL_N_BUCKETS: usize = N_UNIQUE_VALUE_BUCKETS + 1;
